More fixes. Draw text using floating point locations. This will cause all map text to move slightly down and to the right, but will hopefully be a little more accurate when changing resolutions and avoids requiring hacky code to handle integer truncation.

+++ src/nortantis/platform/Painter.java
@@ -27,7 +27,7 @@ public abstract class Painter
 
 	public abstract void setFont(Font font);
 
-	public abstract void drawString(String string, int x, int y);
+	public abstract void drawString(String string, double x, double y);
 
 	public abstract void setTransform(Transform transform);
